nodes/navigator/pyproject.toml
[project]
name = "navigator"
version = "0.0.0"
authors = [{ name = "Your Name", email = "email@email.com" }]
description = "navigator"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "dora-rs >= 0.3.9",
  "scipy>=1.10.1",
]

[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]

[project.scripts]
navigator = "navigator.main:main"

[tool.ruff.lint]
extend-select = [
  "D",   # pydocstyle
  "UP"
]
